Longest Consecutive Sequence
Given an unsorted array of integers nums, return the length of the longest consecutive elements sequence.

You must write an algorithm that runs in O(n) time.

 

Example 1:

Input: nums = [100,4,200,1,3,2]
Output: 4
Explanation: The longest consecutive elements sequence is [1, 2, 3, 4]. Therefore its length is 4.

Example 2:

Input: nums = [0,3,7,2,5,8,4,6,0,1]
Output: 9

 

Constraints:

    0 <= nums.length <= 105
    -109 <= nums[i] <= 109

Idea:

In order to accomplish this task in O(N) time, we'll have to have some way of looking up values, which means a set or map object. The problem with using a set in this situation is that we will also need to prevent redoing the same path over and over again (pushing the time complexity to O(N^2)), which means keeping track of which numbers have been seen.

This would tend to point towards a map approach where we store the numbers in nums as the key and the index of the number as the value in our map (nmap). Then we can set up a seen array based on the indexes of the numbers so that we don't redo already-visited numbers.

(Note: Since we'll be iterating through nums from front to back in the next section, we should store the first index that a number is found at in nmap. Later indexes of duplicate numbers can be ignored.)

But now we run into the issue of potentially finding the middle of a sequence before we find the beginning of the sequence. For this we can take inspiration from a union-find approach, and use seen to store the length of the sequence found when starting at the given number.

We only need to store the path length in the seen index representing the point of entry, because it's the only element that we might later find while iterating through unseen numbers. For the other indexes we pass through while tracking the sequence, we can just fill seen with a 1 to indicate that it has been visited.

Then, if we later find an earlier number in the same sequence, we can notice the number in seen when we reach the visited portion and just add that to our count of numbers.

For example, consider nums = [4,5,6,1,2,3,0]. We start at 4, then track through 5 and 6, filling the seen indexes corresponding to 5 and 6 with a 1 each. Once we reach the end of that chain and have a count of 3, we store that 3 in the seen index corresponding to 4 (seen[0]).

Then, because we've seen 5 and 6, we skip to 1. At 1, we track through 2 and 3, filling them with 1s, then we run into 4, which has a value of 3 stored in seen. At this point, count is 3, but we've just run into another already-discovered chain of 3, so we can fill the seen index corresponding to 1 (seen[3]) with a 6. Then we skip past 2 and 3, and the 0 will lead us back to 1, so we'll have a result of 7 for the seen index corresponding to 0 (seen[6]).

At each step, when we're about to store the count in seen, we can also update our best result so far (ans). Then, once we've reached the end of the iteration, we can just return ans.

    Time Complexity: O(N) where N is the length of nums
    Space Complexity: O(N) for nmap and seen