Minimum Number of Refueling Stops
A car travels from a starting position to a destination which is target miles east of the starting position.

Along the way, there are gas stations.  Each station[i] represents a gas station that is station[i][0] miles east of the starting position, and has station[i][1] liters of gas.

The car starts with an infinite tank of gas, which initially has startFuel liters of fuel in it.  It uses 1 liter of gas per 1 mile that it drives.

When the car reaches a gas station, it may stop and refuel, transferring all the gas from the station into the car.

What is the least number of refueling stops the car must make in order to reach its destination?  If it cannot reach the destination, return -1.

Note that if the car reaches a gas station with 0 fuel left, the car can still refuel there.  If the car reaches the destination with 0 fuel left, it is still considered to have arrived.

 

Example 1:

Input: target = 1, startFuel = 1, stations = []
Output: 0
Explanation: We can reach the target without refueling.

Example 2:

Input: target = 100, startFuel = 1, stations = [[10,100]]
Output: -1
Explanation: We can't reach the target (or even the first gas station).

Example 3:

Input: target = 100, startFuel = 10, stations = [[10,60],[20,30],[30,30],[60,40]]
Output: 2
Explanation: 
We start with 10 liters of fuel.
We drive to position 10, expending 10 liters of fuel.  We refuel from 0 liters to 60 liters of gas.
Then, we drive from position 10 to position 60 (expending 50 liters of fuel),
and refuel from 10 liters to 50 liters of gas.  We then drive to and reach the target.
We made 2 refueling stops along the way, so we return 2.

 

Note:

    1 <= target, startFuel, stations[i][1] <= 10^9
    0 <= stations.length <= 500
    0 < stations[0][0] < stations[1][0] < ... < stations[stations.length-1][0] < target

Solution 1

One of the solution use dynamic programming. Let dp[i][j] be a maximum amount of fuel we can have, when we reached station number i and refueled j times before this station (not included). Let us also add two more points to our stops: starting one and ending one. Then, how we can update dp[i][j]: we can

    Either refuel on previous station, than we have dp[i-1][j-1] minus distance we covered between i-1 and i stations and plus amount of fuel we get at last station.
    Or it can be dp[i-1][j] plus distance we covered between i-1 and i stations.

If we have cand < 0 in our code, than it means we can not reach position i with j refuels, so we leave it equal to minus infinity. Finally, in the end we check last station and find the smallest index where value is not negative and return it.

Complexity

Time complexity is O(n^2), space is O(n^2) as well. Space can be reduced to O(n).


Solution 2

There is also O(n log n) complexity solution, using heaps! Let us follow the following strategy: ride until we can, and if we can not, choose the biggest of stations we met previously and refuel at this station. Why it is working? Because in any case we need to spend target amount of fuel, and if we choose station with biggest amount of fuel inside, then we always better if we choose other station. So, the algorithm will work like this:

    Add starting and ending position to our stations
    Iterate through positions: update fuel: subtract distance we passed
    If it happen that we have negative amount of fuel, we need to refuel at some station previously, we choose the biggest one, increase answer
    If it happen, that we can not reach zero level, it means we are stuck and we can return -1 immediately.
    Push new amount of fuel to our heap for just visited station.

Question
What happens if we have situation like this? -----10-------9--!---, we are at the moment at place ! and we meet two stations previously with fuel 9 recently and with fuel 10 some time ago. Then it seems like it is better to refuel at station with 9. However it is not the case and I spend some time realizing it. If we change from station with 9 to station with 10, when we reach !, we will have more fuel, because we always travel the same distance, but now we have more fuel in our tank.

Complexity

Time complexity is O(n log n), space complexity is O(n).

1
SaraffSunny's avatar
SaraffSunny
846

3 hours ago

61 VIEWS

Idea?

    Maintain a variable max_reachable, which is intialized with startfuel.
    Whenever on reaching a particular station, max_reachable isn't sufficient, try refueling your vehicle.
    From Which gas station, should we refuel?
    Maintain a track of unused gas provided by each station.
    Whenever max_reachable isn't sufficient to reach current station, refuel your vehicle starting with max litres of gas provided by previous stations which are unused.
    Whenever You refuel, increment you answer.
    Whenever after refueling from all previous gas stations, still can't reach current station, return -1.