Maximum Units on a Truck
You are assigned to put some amount of boxes onto one truck. You are given a 2D array boxTypes, where boxTypes[i] = [numberOfBoxesi, numberOfUnitsPerBoxi]:

    numberOfBoxesi is the number of boxes of type i.
    numberOfUnitsPerBoxi is the number of units in each box of the type i.

You are also given an integer truckSize, which is the maximum number of boxes that can be put on the truck. You can choose any boxes to put on the truck as long as the number of boxes does not exceed truckSize.

Return the maximum total number of units that can be put on the truck.

 

Example 1:

Input: boxTypes = [[1,3],[2,2],[3,1]], truckSize = 4
Output: 8
Explanation: There are:
- 1 box of the first type that contains 3 units.
- 2 boxes of the second type that contain 2 units each.
- 3 boxes of the third type that contain 1 unit each.
You can take all the boxes of the first and second types, and one box of the third type.
The total number of units will be = (1 * 3) + (2 * 2) + (1 * 1) = 8.

Example 2:

Input: boxTypes = [[5,10],[2,5],[4,7],[3,9]], truckSize = 10
Output: 91

 

Constraints:

    1 <= boxTypes.length <= 1000
    1 <= numberOfBoxesi, numberOfUnitsPerBoxi <= 1000
    1 <= truckSize <= 106

   Hide Hint #1  
If we have space for at least one box, it's always optimal to put the box with the most units.
   Hide Hint #2  
Sort the box types with the number of units per box non-increasingly.
   Hide Hint #3  
Iterate on the box types and take from each type as many as you can.

Easy problem, that you can solve just sorting all the available boxes by decreasing order of capacity, in order to get the most convenient boxes first.

To do so, we will declare a few support variables:

    res will store how much we can actually fit in our truck;
    currBatch will be used later to store the size of the current batch.

We will then proceed to sort as explained 2 paragraph above and then loop through each element el of boxes (taken as a reference, of course):

    compute currBatch as the minimum between the current truckSize and el[0] (how many boxes of that size are available);
    decrease truckSize accordingly, to reflect loading this batch;
    increase res by currBatch multiplied el[1] (how many units are inside each box of this type);
    if truckSize is now 0, no point in continuing and we can break;


But, wait a moment: a relatively small range and data type, so we know a better way of sorting - doing so in linear time instead of linearithmic, courtesy of the bucket sorting magic (and some extra space to use).

This time we will store info into sizeBucket, an array of 1001 elements (to cover all the provided range) all set to 0 and that we will populate while also storing the extremes of said range in minBucket and maxBucket, increasing each bucket of size boxType[1] by boxType[0] (how many we have) as we go. Be sure to add, not to assign here, since we do not know if we would be given multiple entries of the same size.

Once done, we can loop through the boxes we bucket-sorted going from maxBucket to minBucket (included) and following a logic specular to the previous one.

This bucket sorting version, which despite using buckets, turns out to be even more efficient in terms of space too: