Matchsticks to Square
You are given an integer array matchsticks where matchsticks[i] is the length of the ith matchstick. You want to use all the matchsticks to make one square. You should not break any stick, but you can link them up, and each matchstick must be used exactly one time.

Return true if you can make this square and false otherwise.

 

Example 1:

Input: matchsticks = [1,1,2,2,2]
Output: true
Explanation: You can form a square with length 2, one side of the square came two sticks with length 1.

Example 2:

Input: matchsticks = [3,3,3,3,4]
Output: false
Explanation: You cannot find a way to form a square with all the matchsticks.

 

Constraints:

    1 <= matchsticks.length <= 15
    0 <= matchsticks[i] <= 109

   Hide Hint #1  
Treat the matchsticks as an array. Can we split the array into 4 equal halves?
   Hide Hint #2  
Every matchstick can belong to either of the 4 sides. We don't know which one. Maybe try out all options!
   Hide Hint #3  
For every matchstick, we have to try out each of the 4 options i.e. which side it can belong to. We can make use of recursion for this.
   Hide Hint #4  
We don't really need to keep track of which matchsticks belong to a particular side during recursion. We just need to keep track of the length of each of the 4 sides.
   Hide Hint #5  
When all matchsticks have been used we simply need to see the length of all 4 sides. If they're equal, we have a square on our hands!

Special case of Problem 698 Partition to K Equal Sum Subsets, but here we need to partition into 4 equal sums, we say, put them into buckets (or sides of square).

Please also check my post https://leetcode.com/discuss/general-discussion/1125779/Dynamic-programming-on-subsets-with-examples-explained for more problems which can be solved with this technique.

Denote by dp(mask) possibility to put numbers from bitmask mask into say k subsets, such that sum inside each of the first k-1 subsets is equal to basket . In more detais: we return -1 if this partition is not possible and we return t >= 0, if it is possible, where t is remainder when we divide sum of all used numbers so far by basket: value, of how many we need to put in each group. Given our mask there is several options about what matchstick can be taken the last: the places, where we meet 1 bit. So, if neib is answer for smaller problem with this match removed dfs(mask ^ 1<<j), and neib >= 0, that is solution exists, and neib + nums[j] <= basket, that is we still have place in the last backet, then we return (neib + nums[j]) % basket. We use % here, because basked can be filled fully and in this case we need to start new one.

Complexity

We have O(2^n) states with O(n) transactions for each one, so overall time complexity is (2^n * n) and space complexity is O(2^n).