Number of Matching Subsequences
Given a string s and an array of strings words, return the number of words[i] that is a subsequence of s.

A subsequence of a string is a new string generated from the original string with some characters (can be none) deleted without changing the relative order of the remaining characters.

    For example, "ace" is a subsequence of "abcde".

 

Example 1:

Input: s = "abcde", words = ["a","bb","acd","ace"]
Output: 3
Explanation: There are three strings in words that are a subsequence of s: "a", "acd", "ace".

Example 2:

Input: s = "dsahjpjauf", words = ["ahjpjau","ja","ahbwzgqnuk","tnmlanowax"]
Output: 2

 

Constraints:

    1 <= s.length <= 5 * 104
    1 <= words.length <= 5000
    1 <= words[i].length <= 50
    s and words[i] consist of only lowercase English letters.

We can see that checking if a word is subsequence of s is nothing but ensuring each character of the word occurs in s one after other (i.e, same order). We already have the ordering and indices of all characters in s from start itself.

So, instead of iterating over s each time, we could just iterate over it once and store indices of each character in increasing order. Thereafter, while checking if a word is subsequence, we will just search if there exists an index in s, for each character of given word, that's greater than the index found for previous charcter.

Since we have stored the indices for each character in sorted order, we would just apply binary search over a character's index array instead of linear search. If all characters of given word are found to have corresponding indices in s in an increasing order, we say that word is subsequence of s.